storage: move storage editor entry point from kebab menu to Installation Method
This change makes the storage editor more discoverable, with still being
being presented as a last option when installing.

Resolves: INSTALLER-4655
