+++ src/components/AnacondaWizard.jsx
@@ -18,7 +18,7 @@ import { AnacondaPage } from "./AnacondaPage.jsx";
 import { AnacondaWizardFooter } from "./AnacondaWizardFooter.jsx";
 import { getSteps } from "./steps.js";
 
-export const AnacondaWizard = ({ automatedInstall, currentStepId, dispatch, isFetching, onCritFail, pauseAtSummary, setCurrentStepId, showStorage }) => {
+export const AnacondaWizard = ({ automatedInstall, currentStepId, dispatch, isFetching, onCritFail, pauseAtSummary, setCurrentStepId, setShowStorage, showStorage }) => {
     /**
      * Wizard step page state (reset in `AnacondaWizard` `goToStep` on step change).
      * - **isFormValid** / **setIsFormValid** — Required fields satisfied; reset when the step changes in the wizard.
@@ -43,6 +43,7 @@ export const AnacondaWizard = ({ automatedInstall, currentStepId, dispatch, isFe
         dispatch,
         onCritFail,
         pauseAtSummary,
+        setShowStorage,
     };
 
     const pageContextValue = {
