+++ src/components/storage/cockpit-storage-integration/CockpitStorageIntegration.jsx
@@ -4,30 +4,18 @@
  */
 import cockpit from "cockpit";
 
-import React, { useContext, useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import { Banner } from "@patternfly/react-core/dist/esm/components/Banner/index.js";
 import { Button } from "@patternfly/react-core/dist/esm/components/Button/index.js";
 import { Content } from "@patternfly/react-core/dist/esm/components/Content/index.js";
-import { DropdownItem } from "@patternfly/react-core/dist/esm/components/Dropdown/index.js";
 import { Modal, ModalBody, ModalFooter, ModalHeader, ModalVariant } from "@patternfly/react-core/dist/esm/components/Modal/index.js";
 import { PageSection } from "@patternfly/react-core/dist/esm/components/Page/index.js";
-import { Tooltip } from "@patternfly/react-core/dist/esm/components/Tooltip/index.js";
 import { Flex, FlexItem } from "@patternfly/react-core/dist/esm/layouts/Flex/index.js";
 import { ArrowLeftIcon } from "@patternfly/react-icons/dist/esm/icons/arrow-left-icon";
 import { ExclamationTriangleIcon } from "@patternfly/react-icons/dist/esm/icons/exclamation-triangle-icon";
 import { TimesIcon } from "@patternfly/react-icons/dist/esm/icons/times-icon";
 
-import {
-    getDeviceAncestors
-} from "../../../helpers/storage.js";
-
-import { StorageContext, TargetSystemRootContext } from "../../../contexts/Common.jsx";
-
 import { useMaybeBackdrop } from "../../../hooks/CockpitIntegration.jsx";
-import {
-    useMountPointConstraints,
-    useOriginalDevices,
-} from "../../../hooks/Storage.jsx";
 
 import { CheckStorageDialog } from "./CheckStorageDialog.jsx";
 import { ModifyStorageSideBar } from "./CockpitStorageIntegrationSidebar.jsx";
@@ -168,49 +156,4 @@ export const CockpitStorageIntegration = ({
             </Modal>
         </>
     );
-};
-
-export const ModifyStorage = ({ currentStepId, setShowStorage }) => {
-    const targetSystemRoot = useContext(TargetSystemRootContext);
-    const { diskSelection } = useContext(StorageContext);
-    const devices = useOriginalDevices();
-    const availableDevices = [
-        ...diskSelection.selectedDisks,
-        ...diskSelection.selectedDisks.map(disk => getDeviceAncestors(devices, disk)).flat(),
-    ];
-    const mountPointConstraints = useMountPointConstraints();
-    const isEfi = mountPointConstraints?.some(c => c["required-filesystem-type"]?.v === "efi");
-    const cockpitAnaconda = JSON.stringify({
-        available_devices: availableDevices.map(device => devices[device].path.v),
-        efi: isEfi,
-        mount_point_prefix: targetSystemRoot,
-    });
-    // Allow to modify storage only when we are in the scenario selection page
-    const isAriaDisabled = currentStepId !== "anaconda-screen-method";
-    const item = (
-        <DropdownItem
-          id="modify-storage"
-          isAriaDisabled={isAriaDisabled}
-          onClick={() => {
-              window.sessionStorage.setItem("cockpit_anaconda", cockpitAnaconda);
-              setShowStorage(true);
-          }}
-        >
-            {_("Launch storage editor")}
-        </DropdownItem>
-    );
-
-    if (!isAriaDisabled) {
-        return item;
-    } else {
-        return (
-            <Tooltip
-              id="modify-storage-tooltip"
-              content={_("Storage editor is available only in the `Installation method` step.")}>
-                <span>
-                    {item}
-                </span>
-            </Tooltip>
-        );
-    }
 };
