+++ test/helpers/storage.py
@@ -112,7 +112,6 @@ class StorageDestination():
         self.browser.wait_not_present("#cockpit-storage-integration-check-storage-dialog")
 
     def modify_storage(self):
-        self.browser.click("#toggle-kebab")
         self.browser.click("#modify-storage")
 
     def enter_cockpit_storage(self):
